5328-Update automatic script to deploy Testnet with specific tag from dock…

+++ deploy/testnet/deploy_by_version.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# ./deploy/testnet/deploy.sh <path_to_share_volumes> <>
+#deploy/testnet/deploy_by_version.sh <path_to_share_volumes> <rpc_corsdomain> <deploy_explorer> <version>
 
 localVolumes=$1
 shift
@@ -7,24 +7,26 @@ rpccorsdomain=$1
 shift
 deployExplorer=$1
 shift
+version=$1
+shift
+
+if [ "$localVolumes" == "" ] || [ "$rpccorsdomain" == "" ] || [ "$deployExplorer" == "" ] || [ "$version" == "" ]
+then
+  echo 'Missing params'
+  exit 1
+fi
 
 BASEDIR=$(dirname "$0")
-
-# Remove gev-builder
-sudo docker rmi -f img_builder
-sudo docker rm -f gev-builder
+IMAGE_TAG="registry.gitlab.com/evry/evrynet-client"
 
 # Start builder
-echo "------ Building project ------"
-rm -rf "$BASEDIR"/builder/project
+echo "--- Cleaning executable files ..."
 rm -rf "$BASEDIR"/builder/bin
 rm -rf "$BASEDIR"/bootnode/bin
 rm -rf "$BASEDIR"/nodes/bin/gev
-echo "Cloning evrynet-node from develop branch ..."
-git clone --single-branch -b develop --single-branch git@github.com:Evrynetlabs/evrynet-node.git "$BASEDIR"/builder/project
 
 # Stop all dockers gracefully to avoid DB crash
-echo "Stop docker containers of nodes"
+echo "--- Stop docker containers of nodes"
 "$BASEDIR"/stop_dockers.sh
 
 # Stop bootnode
@@ -33,24 +35,32 @@ sudo docker stop bootnode
 # Clear network bridge
 yes | sudo docker network prune
 
-echo "Building builder container"
-yes | sudo docker-compose -f "$BASEDIR"/docker-compose.yml up -d --force-recreate --build gev-builder
+echo "--- Pulling $IMAGE_TAG image"
+sudo docker pull "$IMAGE_TAG:$version"
+
+echo "--- Checking $IMAGE_TAG:$version image"
+GrepInfo=$(sudo docker images -a | grep "$IMAGE_TAG" | grep "$version" | awk '{print $1}')
+echo "$GrepInfo"
 
 # shellcheck disable=SC2181
-if [ $? -eq 0 ]
+if [ "$GrepInfo" != "" ];
 then
-  rm -rf "$BASEDIR"/builder/project
-  echo "=> Building project successfully!"
+  echo "=> Pulling $IMAGE_TAG:$version image successfully!"
   sleep 3
 
   # Remove old services
+  echo '--- Removing old services'
   sudo docker rmi -f img_bootnode img_node_1 img_node_2 img_node_3
   sudo docker rm -f gev-bootnode gev-node-1 gev-node-2 gev-node-3
 
-  echo "Copy gev, bootnode from gev-builder to ./builder/bin/"
+  echo "--- Creating temporary docker container from $IMAGE_TAG to get bin files"
+  TempBuilderContainer=$(docker create "$IMAGE_TAG:$version")
+
+  echo "--- Copy gev, bootnode from $IMAGE_TAG to ./builder/bin/"
   mkdir "$BASEDIR"/builder/bin/
-  sudo docker cp gev-builder:/evrynet/gev "$BASEDIR"/builder/bin/
-  sudo docker cp gev-builder:/evrynet/bootnode "$BASEDIR"/builder/bin/
+  sudo docker cp "$TempBuilderContainer":/evrynet/gev "$BASEDIR"/builder/bin/
+  sudo docker cp "$TempBuilderContainer":/evrynet/bootnode "$BASEDIR"/builder/bin/
+  docker rm -v "$TempBuilderContainer"
 
   # Start bootnode
   echo "------ Start bootnode ------"
@@ -69,7 +79,7 @@ then
   if [[ "$deployExplorer" == "y" ]]; then
     echo "------ Start explorer ------"
     rm -rf "$BASEDIR"/explorer/web
-    echo "Cloning explorer from master branch ..."
+    echo "--- Cloning explorer from master branch ..."
     git clone git@github.com:Evrynetlabs/explorer.git "$BASEDIR"/explorer/web
 
     sudo docker rmi -f img_explorer
@@ -81,6 +91,6 @@ then
   exit 0
 else
   rm -rf "$BASEDIR"/builder/project
-  echo "Building project failed!"
+  echo "--- Building project failed!"
   exit 1
 fi
